5.0_Assignment/requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
openai